+++ src/main/java/com/zx/sms/codec/sgip12/codec/SgipDeliverResponseMessageCodec.java
@@ -42,7 +42,7 @@ public class SgipDeliverResponseMessageCodec extends MessageToMessageCodec<Messa
 
 	@Override
 	protected void decode(ChannelHandlerContext ctx, Message msg, List<Object> out) throws Exception {
-		long commandId = ((Long) msg.getHeader().getCommandId()).longValue();
+		int commandId =  msg.getHeader().getCommandId();
 		if (packetType.getCommandId() != commandId) {
 			// 不解析，交给下一个codec
 			out.add(msg);
