+++ src/main/java/com/zx/sms/codec/sgip12/codec/SgipSubmitResponseMessageCodec.java
@@ -40,7 +40,7 @@ public class SgipSubmitResponseMessageCodec extends MessageToMessageCodec<Messag
 
 	@Override
 	protected void decode(ChannelHandlerContext ctx, Message msg, List<Object> out) throws Exception {
-		long commandId = ((Long) msg.getHeader().getCommandId()).longValue();
+		int commandId =  msg.getHeader().getCommandId();
 		if (packetType.getCommandId() != commandId) {
 			// 不解析，交给下一个codec
 			out.add(msg);
