+++ src/main/java/com/zx/sms/session/cmpp/SessionStateManager.java
@@ -16,15 +16,15 @@ import com.zx.sms.session.AbstractSessionStateManager;
 /**
  * @author Lihuanghe(18852780@qq.com) 消息发送窗口拜你控制和消息重发 ，消息持久化
  */
-public class SessionStateManager extends AbstractSessionStateManager<Long, Message> {
+public class SessionStateManager extends AbstractSessionStateManager<Integer, Message> {
 	private static final Logger logger = LoggerFactory.getLogger(SessionStateManager.class);
 
-	public SessionStateManager(EndpointEntity entity, ConcurrentMap<Long, VersionObject<Message>> storeMap, boolean preSend) {
+	public SessionStateManager(EndpointEntity entity, ConcurrentMap<Integer, VersionObject<Message>> storeMap, boolean preSend) {
 		super(entity, storeMap, preSend);
 	}
 
 	@Override
-	protected Long getSequenceId(Message msg) {
+	protected Integer getSequenceId(Message msg) {
 		return msg.getHeader().getSequenceId();
 	}
